layers: Decode sFlow's discarded packet event
sFlow has a new specification that allows devices to export discarded
packets and their accompanying drop reasons, defined at
https://sflow.org/sflow_drops.txt

The sFlow decoder has been updated to support decoding this new sample
type. This is based on the existing logic for decoding flow samples, as
a discarded packet event contains the same type of records as a flow
sample.